% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_census_seg.R
\name{all_census_seg}
\alias{all_census_seg}
\title{A wrapper function for get_acs_race and a particular segregation function.}
\usage{
all_census_seg(
  geography = "tract",
  cache_table = FALSE,
  year = 2010,
  key = NULL,
  counties = FALSE,
  seg_func = calc_theilsH,
  mc.cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{geography}{The geography of your data. Defaults to 'tract'.}

\item{cache_table}{Whether or not to cache table names for faster future
access. Defaults to FALSE; if TRUE, only needs to be called once per dataset.
If variables dataset is already cached via the load_variables function, this
can be bypassed.}

\item{year}{The year, or endyear, of the ACS sample or Census. 2012 through
2018 are available for ACS and 2000 and 2010 for Census. Defaults to 2010.}

\item{key}{Your Census API key. Obtain one at
http://api.census.gov/data/key_signup.html}

\item{counties}{Do analysis of all counties rather than metro areas. Default
FALSE.}

\item{seg_func}{a segregation function from the censeg package. Defaults to
`calc_theilsH`}

\item{mc.cores}{The number of cores to use, i.e. at most how many child
processes will be run simultaneously. The option is initialized from
environment variable MC_CORES if set. Must be at least one,
and parallelization requires at least two cores.}
}
\value{
data frame with geography-race specific counts for all
counties or CBSAs.
}
\description{
all_census_seg uses the census api to download race & ethnicity
data from a specified 5 year ACS survey from the table "B03002". From this
table the racial groups White, Black, American Indian or Alaskan Native,
Asian, Native Hawaiian or Pacific Islander, some other race, and two or more
Races are considered and treated as mutually exclusive categories for
individuals who mark Non-Hispanic. Hispanic individuals are designated as
Hispanic no matter what they mark for their Racial identifier and are treated
as a mutually exclusive racial group for the purposes of the analysis. Then,
segregation measures are calculated on the resulting data by either county or
metropolitan area across the US 50 states and DC.
}
\examples{
\dontrun{
# Calculate segregation for all CBSAs from the 2010 census
all_census_seg(year = 2010)
}

}
