% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abclust.R
\name{calc_abclust}
\alias{calc_abclust}
\title{Calculate absolute clustering measure of segregation}
\usage{
calc_abclust(
  group_df,
  out_groups = "NH White",
  parent_geo = c(),
  year = 2010,
  ...
)
}
\arguments{
\item{group_df}{race specific df with a count for every GEOID}

\item{out_groups}{character vector indicating the out group of the
comparison. The default behavior selects "NH White" such that
Non Hispanic White segregation is being evaluated.}

\item{parent_geo}{character vector of additional columns to subdivide analysis}

\item{year}{year to get geography information. Defaults to 2010}

\item{...}{other arguments to pass to geo_distance}
}
\value{
data frame with measures of absolute clustering
}
\description{
calculates absolute clustering on a dataset
}
\examples{
calc_abclust(get_acs_race("DC"))

}
