% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dissimilarity.R
\name{calc_dissimilarity}
\alias{calc_dissimilarity}
\title{Calculate index of dissimilarity measure of segregation}
\usage{
calc_dissimilarity(
  group_df,
  out_groups = "NH White",
  alt_groups = NULL,
  parent_geo = c(),
  ...
)
}
\arguments{
\item{group_df}{race specific df with a count for every GEOID}

\item{out_groups}{character vector indicating the out group of the
comparison. The default behavior selects "NH White" such that
White-Non White segregation is being evaluated.}

\item{alt_groups}{character vector indicating the alt groups of the
comparison. The default behavior selects NULL such that
all groups except the out group are selected.}

\item{parent_geo}{character vector of additional columns to subdivide analysis}

\item{...}{added for compatibility}
}
\value{
single row data frame with index of dissimilarity
}
\description{
calculates index of dissimilarity on a dataset
}
\examples{
calc_dissimilarity(get_acs_race("CA", "Orange", year = 2013))

}
