% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lisa.R
\name{calc_lisa}
\alias{calc_lisa}
\title{Calculate local indicator of spatial association (LISA) measure of segregation}
\usage{
calc_lisa(
  group_df,
  out_groups = "NH White",
  parent_geo = c(),
  year = 2010,
  method = "none",
  ...
)
}
\arguments{
\item{group_df}{race specific df with a count for every GEOID}

\item{out_groups}{character vector indicating the out group of the
comparison. The default behavior selects "NH White" such that
Non Hispanic White clustering is being evaluated.}

\item{parent_geo}{character vector of additional columns to subdivide analysis}

\item{year}{year to get geography information. Defaults to 2010}

\item{method}{character method for p value adjustment}

\item{...}{other arguments to pass to geo_distance}
}
\value{
data frame with measures of lisa local clusters
}
\description{
calculates LISA on a dataset
}
\examples{
calc_lisa(get_decennial_race("DC"))

}
