% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_locquo.R
\name{calc_locquo}
\alias{calc_locquo}
\title{Calculate location quotient measure of segregation}
\usage{
calc_locquo(group_df, out_groups = "NH White", parent_geo = c(), ...)
}
\arguments{
\item{group_df}{race specific df with a count for every GEOID}

\item{out_groups}{character vector indicating the out group of the
comparison. The default behavior selects "NH White" such that
White clustering is being calculated relative the total population.}

\item{parent_geo}{character vector of additional columns to subdivide analysis}

\item{...}{added for compatibility}
}
\value{
data frame with location quotient
}
\description{
calculates location quotient on a dataset
}
\examples{
calc_locquo(get_acs_race("CO", "Boulder", year = 2017))

}
