% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbsa_county_df}
\alias{cbsa_county_df}
\title{Core based statistical areas of the United States county mapping.}
\format{
A data frame with 1808 rows and 6 variables:
\describe{
  \item{cbsacode}{FIPS code for the CBSA}
  \item{cbsattitle}{Common name of the CBSA}
  \item{STATE}{State FIPS code as character with leading zeros}
  \item{COUNTY}{County FIPS code as character with leading zeros}
  \item{statename}{name of state}
  \item{type}{Type of Area (either Metropolitan or Micropolitan)}
}
}
\source{
\url{https://data.nber.org/cbsa-csa-fips-county-crosswalk/cbsa2fipsxw.csv}
}
\usage{
cbsa_county_df
}
\description{
Metropolitan and micropolitan areas of the United States and their associated
counties as designated by OMB 2013 descriptions.
}
\keyword{datasets}
