% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbsa_sf}
\alias{cbsa_sf}
\title{Core based statistical areas of the United States shape file.}
\format{
A data frame with 1808 rows and 6 variables:
\describe{
  \item{cbsacode}{FIPS code for the CBSA}
  \item{cbsattitle}{Common name of the CBSA}
  \item{type}{Type of Area (either Metropolitan or Micropolitan)}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\usage{
cbsa_sf
}
\description{
Metropolitan and micropolitan areas of the United States and their associated
shape geometries for mapping.
}
\keyword{datasets}
