% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_distance.R
\name{geo_distance}
\alias{geo_distance}
\title{Calculate a distance matrix from an sf object}
\usage{
geo_distance(
  GEOIDs = NULL,
  year = 2018,
  geoid_sf = NULL,
  matrix = FALSE,
  adjust = TRUE
)
}
\arguments{
\item{GEOIDs}{Character vector of GEOIDs of either tracts or block groups}

\item{year}{The year, or endyear, of the ACS sample or Census. 2012 through
2018 are available for ACS and 2000 and 2010 for Census. Defaults to 2010.}

\item{geoid_sf}{Alternative specification where an sf object is supplied and
distances are calculated directly from that object}

\item{matrix}{Logical whether to return a matrix rather than data frame.}

\item{adjust}{Logical, adjust using Massey Denton specification.}
}
\value{
data frame or matrix indicating distance between geometry centroids
in kilometers.
}
\description{
Calculates a distance matrix from a list of geoids which
correspond to either tract of block group ids for a given year. Alternatively
calculates distance matrix for a sf data frame of polygons.
}
\examples{
# get distances of all block groups in DC

geo_distance(geoid_sf = block_groups("DC", class = "sf"), matrix = TRUE)

}
