% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_metro.R
\name{get_acs_metro}
\alias{get_acs_metro}
\title{A wrapper function of get_acs which gets data for 50 states plus DC}
\usage{
get_acs_metro(
  geography,
  variables = NULL,
  table = NULL,
  cache_table = FALSE,
  year = 2018,
  output = "tidy",
  summary_var = NULL,
  key = NULL,
  moe_level = 90,
  survey = "acs5",
  show_call = FALSE,
  ...
)
}
\arguments{
\item{geography}{unit for which to pull data}

\item{variables}{Character string or vector of character strings of
variable IDs.}

\item{table}{The Census table for which you would like to request all
variables. Uses lookup tables to identify the variables; performs faster
when variable table already exists through load_variables(cache = TRUE).
Only one table may be requested per call.}

\item{cache_table}{Whether or not to cache table names for faster future
access. Defaults to FALSE; if TRUE, only needs to be called once per dataset.
If variables dataset is already cached via the load_variables function,
this can be bypassed.}

\item{year}{The year for which you are requesting data. Default 2018.}

\item{output}{One of "tidy" (the default) in which each row represents an
enumeration unit-variable combination, or "wide" in which each row
represents an enumeration unit and the variables are in the columns.}

\item{summary_var}{Character string of a "summary variable" from the
decennial Census to be included in your output. Usually a variable
(e.g. total population) that you'll want to use as a denominator.}

\item{key}{Your Census API key. Obtain one at
http://api.census.gov/data/key_signup.html}

\item{moe_level}{The confidence level of the returned margin of error.
One of 90 (the default), 95, or 99.}

\item{survey}{The ACS contains one-year, three-year, and five-year surveys
expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}

\item{show_call}{if TRUE, display call made to Census API. This can be very
useful in debugging and determining if error messages returned are due to
tidycensus or the Census API. Copy to the API call into a browser and see
what is returned by the API directly. Defaults to FALSE.}

\item{...}{Other argumenets to pass to get_decennial.}
}
\value{
data frame with census info and adjoined cbsa location
}
\description{
get_acs_metro uses the census api to download data for all
50 states plus DC for "on spine" geographies county and below. This method
will associate those geographies with a CBSA as well.
}
\examples{
\dontrun{
# get Guat population for all tracts in US with appropriate metro info
get_acs_metro("tract", c(Guatemalan = "B03001_010"))
}

}
