% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_theilsH.R
\name{calc_theilsH}
\alias{calc_theilsH}
\title{Calculate entropy and Theil's-H measure of segregation}
\usage{
calc_theilsH(group_tract_df, ...)
}
\arguments{
\item{group_tract_df}{race specific df with a count for every census tract}

\item{...}{added for compatibility}
}
\value{
single row data frame with measures of Theil's-H and entropy
}
\description{
calculates Theil's-H and Entropy on a dataset
}
\examples{
calc_theilsH(get_acs_race("WA", "King"))

}
